+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export * from './typeaheadTags/typeaheadTags.component';
+export * from './typeaheadTags/typeaheadTagsControlValueAccessor.directive';