+++ src/modules/bootstrap.module.ts
@@ -8,6 +8,7 @@ import {BOOTSTRAP_SELECT_DIRECTIVES} from '../directives/bootstrapSelect';
 import {TYPEAHEAD_DIRECTIVES} from '../directives/typeahead';
 import {BootstrapPopoverDirective, BootstrapTooltipDirective} from '../directives/misc';
 import {BootstrapSwitchDirective} from '../directives/bootstrapSwitch/bootstrapSwitch.directive';
+import {TypeaheadTagsComponent, TypeaheadTagsControlValueAccessor} from '../components';
 
 /**
  * Module for bootstrap components
@@ -15,8 +16,32 @@ import {BootstrapSwitchDirective} from '../directives/bootstrapSwitch/bootstrapS
 @NgModule(
 {
     imports: [CommonModule],
-    declarations: [DialogComponent, ConfirmationDialogComponent, DATETIME_PICKER_DIRECTIVES, BOOTSTRAP_SELECT_DIRECTIVES, TYPEAHEAD_DIRECTIVES, BootstrapPopoverDirective, BootstrapTooltipDirective, BootstrapSwitchDirective],
-    exports: [DialogComponent, ConfirmationDialogComponent, DATETIME_PICKER_DIRECTIVES, BOOTSTRAP_SELECT_DIRECTIVES, TYPEAHEAD_DIRECTIVES, BootstrapPopoverDirective, BootstrapTooltipDirective, BootstrapSwitchDirective]
+    declarations: 
+    [
+        DialogComponent,
+        ConfirmationDialogComponent,
+        DATETIME_PICKER_DIRECTIVES,
+        BOOTSTRAP_SELECT_DIRECTIVES,
+        TYPEAHEAD_DIRECTIVES,
+        BootstrapPopoverDirective,
+        BootstrapTooltipDirective,
+        BootstrapSwitchDirective,
+        TypeaheadTagsComponent,
+        TypeaheadTagsControlValueAccessor
+    ],
+    exports: 
+    [
+        DialogComponent,
+        ConfirmationDialogComponent,
+        DATETIME_PICKER_DIRECTIVES,
+        BOOTSTRAP_SELECT_DIRECTIVES,
+        TYPEAHEAD_DIRECTIVES,
+        BootstrapPopoverDirective,
+        BootstrapTooltipDirective,
+        BootstrapSwitchDirective,
+        TypeaheadTagsComponent,
+        TypeaheadTagsControlValueAccessor
+    ]
 })
 export class BootstrapModule
 {