+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Version 4.0.4
+ - added new component `TypeaheadTagsComponent`
+
 ## Version 4.0.3
  - adding css class to input if no value is selected and text is inserted in input
  - support for using free text as value of typeahead